requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
python-dotenv==1.0.1
feedparser==6.0.11
readability-lxml==0.8.1
openai==1.50.2
apscheduler==3.10.4
slowapi==0.1.9
